requirements.txt
langchain
langchain-community
langchain-core
python-dotenv
langchain-google-genai
google-generativeai
pandas
faiss-cpu
streamlit
pytest